+++ xclim/core/utils.py
@@ -9,6 +9,7 @@ Helper functions for the indices computation, indicator construction and other t
 from collections import defaultdict
 from enum import IntEnum
 from functools import partial
+from importlib.resources import open_text
 from inspect import Parameter
 from types import FunctionType
 from typing import Callable, NewType, Optional, Sequence, Union
@@ -19,12 +20,16 @@ import xarray as xr
 from boltons.funcutils import update_wrapper
 from dask import array as dsk
 from xarray import DataArray, Dataset
+from yaml import safe_load
 
 #: Type annotation for strings representing full dates (YYYY-MM-DD), may include time.
 DateStr = NewType("DateStr", str)
 
 #: Type annotation for strings representing dates without a year (MM-DD).
 DayOfYearStr = NewType("DayOfYearStr", str)
+
+# Official variables definitions
+variables = safe_load(open_text("xclim.data", "variables.yml"))["variables"]
 
 
 def wrapped_partial(
@@ -65,6 +70,11 @@ def wrapped_partial(
     fully_wrapped = update_wrapper(
         partial_func, func, injected=list(fixed.keys()), hide_wrapped=True
     )
+
+    # Store all injected params,
+    injected = getattr(func, "_injected", {})
+    injected.update(fixed)
+    fully_wrapped._injected = injected
     return fully_wrapped